Abstracts Admin priviledge checking and Book adding and updating logic to helper functions
